sports ministry: parse out 3d cart "option" values from report
This helper script takes a CSV export of sales from the 3d cart store
and parsed out the "options" from the itemname column and put them
into a new CSV consisting of just these values.

Specifically, when someone registers for a sport on the 3d cart store,
there's a series of questions that must be answered before the item
(i.e., registration) can be added to the cart, such as: athlete name,
uniform size, ...etc.  All of these questions get mushed into a single
column in the CSV sales report that you can download from the store,
making it fairly unintelligible for a human to read.  Plus, the sports
ministry doesn't need all the billing information that comes in the
CSV sales report -- all they need is the answers to the questions that
accompanied the purchase of the sports registration.  This script
simply parses out those answers and outputs them into a new CSV that
can be easily used by the Sports Ministry to list all the athletes who
have registered, etc.

The script is currently hard-coded to look for item ID's with
"2015-SPR" in the name.  This should probably be turned into a command
line option.