pto flower sale: parse out 3d cart "option" values from report
This helper script takes a CSV export of sales from the 3d cart store
and parsed out the "options" from the itemname column and add extra
rows at the end with those options (in this case, the only option is
the color of some of the flowers).

The script is currently hard-coded to look for item ID's with
"1368.81" in the name.  This should probably be turned into a command
line option.